Add Abilities, Move Caster to it's own file, move subclasses to their own file and test them
